Add WMFTestAppEditCounter
Adds a testing Counter implementation for counting app edits. This further
serves as an example of an edit counter with qualifications.

On edit success, the request headers are examined to determine whether the
edit is coming from a 'WikipediaApp/' user agent.  (This is the same
strategy used in the MobileApp extension to determine whether to apply a
'mobile app edit' tag.)

On revert, the tag summary is examined for the presence of the substring
'app edit'.

The dummy conditional checking logic is removed from the abstract counter
class.  Any conditional logic should live in the various concrete Counter
implementations.